+++ modulo5/classes-encapsulamento/package.json
@@ -23,6 +23,7 @@
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
     "knex": "^2.1.0",
-    "mysql": "^2.18.1"
+    "mysql": "^2.18.1",
+    "ts-node": "^10.8.2"
   }
 }
