+++ modulo5/classes-encapsulamento/src/Transaction.ts
@@ -1,0 +1,30 @@
+// type Transaction = {
+//     description: string,
+//     value: number,
+//     date: string
+//   }
+
+export class Transaction {
+    private description: string
+    private value: number
+    private date: string
+
+    constructor(description: string, value: number, date: string) {
+            this.description = description,
+            this.value = value,
+            this.date = date
+    }
+
+    public getDescription():string {
+        return this.description
+    }
+
+    public getValue():number {
+        return this.value
+    }
+
+    public getDate():string {
+        return this.date
+    }
+}
+
