--- modulo5/classes-encapsulamento/src/data/UserAccount.ts
@@ -1,21 +1,0 @@
-import { Transaction } from "./Transaction";
-
-class UserAccount {
-    private cpf: string;
-    private name: string;
-    private age: number;
-    private balance: number = 0;
-    private transactions: Transaction[] = [];
-  
-    constructor(
-       cpf: string,
-       name: string,
-       age: number,
-    ) {
-       console.log("Chamando o construtor da classe UserAccount")
-       this.cpf = cpf;
-       this.name = name;
-       this.age = age;
-    }
-  
-  }
